July 31st Updates
Replacing BMP files that got exported as LMP to PNG so that this problem no longer persists, also worked with GaX on Devestator code abilities, as well as some adjustments to the ability code in general. However going forward I need to make more adjustments to the ability code, will touch base on Water with that soon.